"Refactor package name to be discover instead of "discovery" and also fixed the map with the center button available again!"
