+++ app/src/main/java/com/lastaoutdoor/lasta/ui/navigation/MenuNavGraph.kt
@@ -10,7 +10,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.lastaoutdoor.lasta.R
 import com.lastaoutdoor.lasta.ui.screen.activities.MoreInfoScreen
-import com.lastaoutdoor.lasta.ui.screen.discovery.DiscoveryScreen
+import com.lastaoutdoor.lasta.ui.screen.discover.DiscoveryScreen
 import com.lastaoutdoor.lasta.ui.screen.favorites.FavoritesScreen
 import com.lastaoutdoor.lasta.ui.screen.profile.ProfileScreen
 import com.lastaoutdoor.lasta.ui.screen.social.AddFriendScreen
