+++ app/src/main/java/com/lastaoutdoor/lasta/ui/screen/discover/DiscoveryScreen.kt
@@ -1,4 +1,4 @@
-package com.lastaoutdoor.lasta.ui.screen.discovery
+package com.lastaoutdoor.lasta.ui.screen.discover
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -36,7 +37,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
@@ -49,8 +49,8 @@ import com.lastaoutdoor.lasta.ui.components.DisplaySelection
 import com.lastaoutdoor.lasta.ui.components.SearchBarComponent
 import com.lastaoutdoor.lasta.ui.components.SeparatorComponent
 import com.lastaoutdoor.lasta.ui.navigation.LeafScreen
-import com.lastaoutdoor.lasta.ui.screen.discovery.components.ModalUpperSheet
-import com.lastaoutdoor.lasta.ui.screen.discovery.components.RangeSearchComposable
+import com.lastaoutdoor.lasta.ui.screen.discover.components.ModalUpperSheet
+import com.lastaoutdoor.lasta.ui.screen.discover.components.RangeSearchComposable
 import com.lastaoutdoor.lasta.ui.screen.map.MapScreen
 import com.lastaoutdoor.lasta.viewmodel.DiscoveryScreenType
 import com.lastaoutdoor.lasta.viewmodel.DiscoveryScreenViewModel
@@ -74,18 +74,21 @@ fun DiscoveryScreen(
   if (screen == DiscoveryScreenType.LIST) {
     LazyColumn(
         modifier =
-            Modifier.testTag("discoveryScreen").background(MaterialTheme.colorScheme.background)) {
+        Modifier
+            .testTag("discoveryScreen")
+            .background(MaterialTheme.colorScheme.background)) {
           item { HeaderComposable(updatePopup = { isRangePopup = true }) }
 
           item {
-            SeparatorComponent() // Add a separator between the header and the activities
             Spacer(modifier = Modifier.height(8.dp))
             ActivitiesDisplay(navController, moreInfoScreenViewModel = moreInfoScreenViewModel)
           }
         }
   } else if (screen == DiscoveryScreenType.MAP) {
-    MapScreen()
-    HeaderComposable(updatePopup = { isRangePopup = true })
+      Column {
+          HeaderComposable(updatePopup = { isRangePopup = true })
+          Box(modifier = Modifier.fillMaxHeight()) { MapScreen() }
+      }
   }
 
   // Add the modal upper sheet
@@ -103,13 +106,15 @@ fun HeaderComposable(
   val iconSize = 48.dp // Adjust icon size as needed
 
   Surface(
-      modifier = Modifier.fillMaxWidth().graphicsLayer { alpha = 0.9f },
+      modifier = Modifier.fillMaxWidth(),
       color = MaterialTheme.colorScheme.background,
       shadowElevation = 3.dp) {
         Column {
           // Location bar
           Row(
-              modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp, vertical = 8.dp),
+              modifier = Modifier
+                  .fillMaxWidth()
+                  .padding(horizontal = 16.dp, vertical = 8.dp),
               verticalAlignment = Alignment.CenterVertically) {
                 Column {
                   Row {
@@ -117,7 +122,9 @@ fun HeaderComposable(
 
                     IconButton(
                         onClick = updatePopup,
-                        modifier = Modifier.size(24.dp).testTag("listSearchOptionsEnableButton")) {
+                        modifier = Modifier
+                            .size(24.dp)
+                            .testTag("listSearchOptionsEnableButton")) {
                           Icon(
                               Icons.Outlined.KeyboardArrowDown,
                               contentDescription = "Filter",
@@ -134,7 +141,9 @@ fun HeaderComposable(
 
           // Search bar with toggle buttons
           Row(
-              modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp, vertical = 8.dp),
+              modifier = Modifier
+                  .fillMaxWidth()
+                  .padding(horizontal = 16.dp, vertical = 8.dp),
               verticalAlignment = Alignment.CenterVertically) {
                 SearchBarComponent(Modifier.weight(1f), onSearch = { /*TODO*/})
                 Spacer(modifier = Modifier.width(8.dp))
@@ -146,7 +155,9 @@ fun HeaderComposable(
                 }
               }
           Row(
-              modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp, vertical = 8.dp),
+              modifier = Modifier
+                  .fillMaxWidth()
+                  .padding(horizontal = 16.dp, vertical = 8.dp),
               verticalAlignment = Alignment.CenterVertically,
               horizontalArrangement = Arrangement.Center) {
                 val contex = LocalContext.current
@@ -160,7 +171,9 @@ fun HeaderComposable(
 
           if (screen == DiscoveryScreenType.LIST) {
             Row(
-                modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp, vertical = 8.dp),
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(horizontal = 16.dp, vertical = 8.dp),
                 verticalAlignment = Alignment.CenterVertically) {
                   Text(
                       LocalContext.current.getString(R.string.filter_by),
@@ -194,27 +207,31 @@ fun ActivitiesDisplay(
   for (a in activities) {
     Card(
         modifier =
-            Modifier.fillMaxWidth()
-                .wrapContentHeight()
-                .padding(vertical = 8.dp, horizontal = 16.dp)
-                .clickable(
-                    onClick = {
-                      moreInfoScreenViewModel.changeActivityToDisplay(a)
-                      navController.navigate(LeafScreen.MoreInfo.route)
-                    }),
+        Modifier
+            .fillMaxWidth()
+            .wrapContentHeight()
+            .padding(vertical = 8.dp, horizontal = 16.dp)
+            .clickable(
+                onClick = {
+                    moreInfoScreenViewModel.changeActivityToDisplay(a)
+                    navController.navigate(LeafScreen.MoreInfo.route)
+                }),
         shape = RoundedCornerShape(8.dp),
         elevation = CardDefaults.cardElevation(defaultElevation = 6.dp),
     ) {
       Column {
         Row(
-            modifier = Modifier.fillMaxWidth().padding(8.dp),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp),
             verticalAlignment = Alignment.CenterVertically,
             horizontalArrangement = Arrangement.SpaceBetween) {
               Box(
                   modifier =
-                      Modifier.shadow(4.dp, RoundedCornerShape(30))
-                          .background(MaterialTheme.colorScheme.primary, RoundedCornerShape(10.dp))
-                          .padding(PaddingValues(8.dp))) {
+                  Modifier
+                      .shadow(4.dp, RoundedCornerShape(30))
+                      .background(MaterialTheme.colorScheme.primary, RoundedCornerShape(10.dp))
+                      .padding(PaddingValues(8.dp))) {
                     Text(
                         text = LocalContext.current.getString(R.string.Climbing),
                         style = MaterialTheme.typography.labelMedium,
@@ -233,7 +250,9 @@ fun ActivitiesDisplay(
 
         Spacer(modifier = Modifier.height(16.dp))
         Row(
-            modifier = Modifier.fillMaxWidth().padding(8.dp),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp),
             verticalAlignment = Alignment.CenterVertically) {
               Text(
                   text = a.locationName ?: "Unnamed Activity",
@@ -242,7 +261,9 @@ fun ActivitiesDisplay(
             }
         SeparatorComponent()
         Row(
-            modifier = Modifier.fillMaxWidth().padding(8.dp),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp),
             verticalAlignment = Alignment.CenterVertically) {
               Icon(
                   imageVector = Icons.Default.Star,
