+++ app/src/main/java/com/lastaoutdoor/lasta/ui/screen/discover/components/LocalitySelectionDropdown.kt
@@ -1,4 +1,4 @@
-package com.lastaoutdoor.lasta.ui.screen.discovery.components
+package com.lastaoutdoor.lasta.ui.screen.discover.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
